DataLabellingUI/requirements.txt
sqlalchemy==1.3.20
psycopg2-binary==2.8.6
streamlit==0.69.2
pillow==8.0.1
pandas==1.1.2

